+++ Runtime/ShaderStructs/LightSource.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 
 public struct LightSource {
-  public const int size = (3 + 3 + 3 + 3 + 1 + 1 + 1) * 4;
+  public const int size = (3 + 3 + 3 + 3 + 1 + 1 + 1 + 1) * 4;
 
   public Vector3 color; // 3 * 4 bytes
   public Vector3 direction; // 3 * 4 bytes
@@ -9,6 +9,7 @@ public struct LightSource {
   public Vector3 worldPosition; // 3 * 4 bytes
   public float range; // 1 * 4 bytes
   public float spotCos; // 1 * 4 bytes
+  public float spotFactor; // 1 * 4 bytes
   public uint type; // 1 * 4 bytes
 
   public LightSource(UnityEngine.Experimental.Rendering.VisibleLight light, Matrix4x4 worldToVoxel) {
@@ -17,7 +18,18 @@ public struct LightSource {
     worldPosition = light.localToWorld.GetColumn(3);
     voxelPosition = worldToVoxel * new Vector4(worldPosition.x, worldPosition.y, worldPosition.z, 1f);
     range = light.range;
-    spotCos = light.lightType == LightType.Spot ? Mathf.Cos(.5f * light.spotAngle * Mathf.Deg2Rad) : -1f;
+    spotCos = 0f;
+    spotFactor = 0f;
     type = (uint)light.lightType;
+
+    if (light.lightType == LightType.Spot) {
+      float halfSpotRadian = .5f * light.spotAngle * Mathf.Deg2Rad;
+
+      spotCos = Mathf.Cos(halfSpotRadian);
+
+      if (light.spotAngle > 0f) {
+        spotFactor = 1f / (Mathf.Cos(Mathf.Atan(Mathf.Tan(halfSpotRadian) * 46f / 64f)) - spotCos);
+      }
+    }
   }
 }